requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
markdown==3.7
pydantic==2.9.2
pydantic[email]==2.9.2
pydantic-extra-types==2.9.0
phonenumbers==8.13.45
rendercv-fonts==0.5.1
ruamel.yaml==0.18.6
typst==0.14.0
